Логотип + рефактор хедера
Добавил загрузку домашней страницы при клике на логотип
Создал папку  nav-bar в которой лежит все, что связано с навигацей по страничке
Сделан рефактор кода
