+++ src/index.js
@@ -3,4 +3,4 @@ import './js/search';
 import './js/library/library-filters';
 import './js/modal/modal';
 import './js/pagination';
-import './js/nav-btns-classes';
+import './js/nav-bar/nav-bar';
