+++ src/js/refs.js
@@ -1,10 +1,10 @@
 export const refs = {
   //-----nav-menu-refs------
-  
   searchInput: document.querySelector('.search-input'),
   homeBtn: document.querySelector('.home-page'),
   libraryBtn: document.querySelector('.library-page'),
   gallery: document.querySelector('.gallery__list'),
+  logo: document.querySelector('.page-header__logo'),
   //-----library-refs-----
   libButtons: document.querySelector('.my-library'),
   queueBtn: document.querySelector('.queue'),
